+++ Synthesizer/Oscillators/Base Implementations/Oscillator.cpp
@@ -5,7 +5,7 @@ Oscillator::Oscillator() {
     phase_increment = calculate_phase_increment(1.0, 44100.0);
 }
 
-Oscillator::Oscillator(double hz, double sample_rate) {
+Oscillator::Oscillator(float hz, float sample_rate) {
     current_phase = 0;
     phase_increment = calculate_phase_increment(hz, sample_rate);
 }
@@ -17,6 +17,6 @@ Oscillator::Oscillator(const Oscillator& other) {
 
 Oscillator::~Oscillator() = default;
 
-void Oscillator::set_frequency(double hz, double sample_rate) {
+void Oscillator::set_frequency(float hz, float sample_rate) {
     phase_increment = calculate_phase_increment(hz, sample_rate);
 }
