+++ Synthesizer/Oscillators/Base Implementations/SawtoothOscillator.cpp
@@ -3,12 +3,12 @@
 
 SawtoothOscillator::SawtoothOscillator() = default;
 
-SawtoothOscillator::SawtoothOscillator(double hz, double sample_rate)
+SawtoothOscillator::SawtoothOscillator(float hz, float sample_rate)
     : Oscillator(hz, sample_rate) {
 }
 
 SawtoothOscillator::SawtoothOscillator(const SawtoothOscillator& other) = default;
 
-double SawtoothOscillator::get_sample() {
+float SawtoothOscillator::get_sample() {
     return std::fmod(current_phase, TWO_PI);
 }
